Refactor code and test kata-like config
Refactor the code so that we can add new
configurations without deleting the old ones.

There is still some work to be done to make this
cleaner.

I am also picking up a number of changes that I
did not properly commit. Mainly these add support
for more complex guests with more devices.

The hope was that I could track down the latency
that Carlos ran into using SEV with Kata,
but I was not able to reproduce.

Signed-off-by: Tobin Feldman-Fitzthum <tobin@ibm.com>